pyproject.toml
[tool.poetry]
name = "nba-fantasy-optimize"
version = "0.2.0"
description = ""
authors = ["Sharad Vikram <sharad.vikram@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.2.4"
tqdm = "^4.60.0"
click = "^7.1.2"
matplotlib = "^3.4.1"
seaborn = "^0.11.1"
numpy = "^1.20.2"
beautifulsoup4 = "^4.9.3"
tabulate = "^0.8.9"
pyfiglet = "^0.8.post1"
yaspin = "^1.5.0"
yfpy = "^4.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.7"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
